Update sim900.h
Hi,
I had a problem with the library because though the SIM is in the original country it was in roaming due to agreements with others operators in case there is no signal and anyway in Europe there are no costs for roaming so someone may want to use SIM card in another country. For example I'm building an antitheft system for my car and I need it to work in every country. 

Set ROAMING to true to activate it